requirements.txt
# Kafka dependencies
kafka-python==2.0.2
confluent-kafka==2.2.0

# Spark dependencies
pyspark==3.5.0
py4j==0.10.9.7

# Data processing
numpy==1.24.3
pandas==2.0.3
pyarrow==13.0.0

# Monitoring
prometheus-client==0.18.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
click==8.1.7
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
